@@ -1,28 +1,50 @@
 module github.com/knqyf263/pet
 
-go 1.16
+go 1.20
 
 require (
 	github.com/BurntSushi/toml v0.3.0
+	github.com/atotto/clipboard v0.1.4
 	github.com/briandowns/spinner v0.0.0-20170614154858-48dbb65d7bd5
+	github.com/chzyer/logex v1.1.10 // indirect
 	github.com/chzyer/readline v0.0.0-20180603132655-2972be24d48e
+	github.com/chzyer/test v0.0.0-20210722231415-061457976a23 // indirect
 	github.com/fatih/color v1.7.0
-	github.com/golang/protobuf v1.1.0 // indirect
 	github.com/google/go-github v15.0.0+incompatible
-	github.com/google/go-querystring v0.0.0-20170111101155-53e6ce116135 // indirect
 	github.com/inconshreveable/mousetrap v1.0.0 // indirect
-	github.com/jroimartin/gocui v0.4.0
-	github.com/mattn/go-colorable v0.0.9 // indirect
-	github.com/mattn/go-isatty v0.0.3 // indirect
-	github.com/mattn/go-runewidth v0.0.2
-	github.com/nsf/termbox-go v0.0.0-20180509163535-21a4d435a862 // indirect
+	github.com/mattn/go-runewidth v0.0.10
 	github.com/pkg/errors v0.8.0
 	github.com/spf13/cobra v0.0.3
 	github.com/spf13/pflag v1.0.1 // indirect
-	github.com/xanzy/go-gitlab v0.10.5
-	golang.org/x/crypto v0.0.0-20180608092829-8ac0e0d97ce4
-	golang.org/x/net v0.0.0-20180530234432-1e491301e022 // indirect
-	golang.org/x/oauth2 v0.0.0-20180603041954-1e0a3fa8ba9a
-	golang.org/x/sys v0.0.0-20191026070338-33540a1f6037 // indirect
-	google.golang.org/appengine v1.0.0 // indirect
+	github.com/xanzy/go-gitlab v0.50.3
+	//github.com/xanzy/go-gitlab v0.10.5
+	golang.org/x/crypto v0.17.0
+	golang.org/x/oauth2 v0.0.0-20181106182150-f42d05182288
+	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c // indirect
+	gopkg.in/alessio/shellescape.v1 v1.0.0-20170105083845-52074bc9df61
+)
+
+require (
+	github.com/awesome-gocui/gocui v1.1.0
+	github.com/go-test/deep v1.1.0
+)
+
+require (
+	github.com/alessio/shellescape v1.4.1 // indirect
+	github.com/gdamore/encoding v1.0.0 // indirect
+	github.com/gdamore/tcell/v2 v2.4.0 // indirect
+	github.com/golang/protobuf v1.2.0 // indirect
+	github.com/google/go-querystring v1.0.0 // indirect
+	github.com/hashicorp/go-cleanhttp v0.5.1 // indirect
+	github.com/hashicorp/go-retryablehttp v0.6.8 // indirect
+	github.com/lucasb-eyer/go-colorful v1.0.3 // indirect
+	github.com/mattn/go-colorable v0.0.9 // indirect
+	github.com/mattn/go-isatty v0.0.3 // indirect
+	github.com/rivo/uniseg v0.1.0 // indirect
+	golang.org/x/net v0.17.0 // indirect
+	golang.org/x/sys v0.15.0 // indirect
+	golang.org/x/term v0.15.0 // indirect
+	golang.org/x/text v0.14.0 // indirect
+	golang.org/x/time v0.0.0-20191024005414-555d28b269f0 // indirect
+	google.golang.org/appengine v1.3.0 // indirect
 )
