@@ -7,9 +7,9 @@ import (
 	"strings"
 	"syscall"
 
-	"github.com/fatih/color"
 	"github.com/knqyf263/pet/config"
 	"github.com/spf13/cobra"
+	"gopkg.in/alessio/shellescape.v1"
 )
 
 // execCmd represents the exec command
@@ -25,22 +25,19 @@ func execute(cmd *cobra.Command, args []string) (err error) {
 
 	var options []string
 	if flag.Query != "" {
-		options = append(options, fmt.Sprintf("--query %s", flag.Query))
+		options = append(options, fmt.Sprintf("--query %s", shellescape.Quote(flag.Query)))
 	}
 
-	commands, err := filter(options)
+	commands, err := filter(options, flag.FilterTag)
 	if err != nil {
 		return err
 	}
 	command := strings.Join(commands, "; ")
-	if config.Flag.Debug {
-		fmt.Printf("Command: %s\n", command)
-	}
-	if config.Flag.Command {
-		fmt.Printf("%s: %s\n", color.YellowString("Command"), command)
-	}
 
+	// Show final command before executing it
+	fmt.Printf("> %s\n", command)
 	signal.Ignore(syscall.SIGINT)
+
 	return run(command, os.Stdin, os.Stdout)
 }
 
@@ -50,6 +47,6 @@ func init() {
 		`Enable colorized output (only fzf)`)
 	execCmd.Flags().StringVarP(&config.Flag.Query, "query", "q", "",
 		`Initial value for query`)
-	execCmd.Flags().BoolVarP(&config.Flag.Command, "command", "c", false,
-		`Show the command with the plain text before executing`)
+	execCmd.Flags().StringVarP(&config.Flag.FilterTag, "tag", "t", "",
+		`Filter tag`)
 }
