@@ -16,19 +16,20 @@ var Conf Config
 
 // Config is a struct of config
 type Config struct {
-	General GeneralConfig
-	Gist    GistConfig
-	GitLab  GitLabConfig
+	General GeneralConfig `toml:"General"`
+	Gist    GistConfig    `toml:"Gist"`
+	GitLab  GitLabConfig  `toml:"GitLab"`
 }
 
 // GeneralConfig is a struct of general config
 type GeneralConfig struct {
-	SnippetFile string `toml:"snippetfile"`
-	Editor      string `toml:"editor"`
-	Column      int    `toml:"column"`
-	SelectCmd   string `toml:"selectcmd"`
-	Backend     string `toml:"backend"`
-	SortBy      string `toml:"sortby"`
+	SnippetFile string   `toml:"snippetfile"`
+	Editor      string   `toml:"editor"`
+	Column      int      `toml:"column"`
+	SelectCmd   string   `toml:"selectcmd"`
+	Backend     string   `toml:"backend"`
+	SortBy      string   `toml:"sortby"`
+	Cmd         []string `toml:"cmd"`
 }
 
 // GistConfig is a struct of config for Gist
@@ -48,6 +49,7 @@ type GitLabConfig struct {
 	ID          string `toml:"id"`
 	Visibility  string `toml:"visibility"`
 	AutoSync    bool   `toml:"auto_sync"`
+	Insecure    bool   `toml:"skip_ssl"`
 }
 
 // Flag is global flag variable
@@ -57,6 +59,7 @@ var Flag FlagConfig
 type FlagConfig struct {
 	Debug     bool
 	Query     string
+	FilterTag string
 	Command   bool
 	Delimiter string
 	OneLine   bool
@@ -103,7 +106,7 @@ func (cfg *Config) Load(file string) error {
 		}
 	}
 	cfg.General.Column = 40
-	cfg.General.SelectCmd = "fzf"
+	cfg.General.SelectCmd = "fzf --ansi --layout=reverse --border --height=90% --pointer=* --cycle --prompt=Snippets:"
 	cfg.General.Backend = "gist"
 
 	cfg.Gist.FileName = "pet-snippet.toml"
@@ -116,7 +119,9 @@ func (cfg *Config) Load(file string) error {
 
 // GetDefaultConfigDir returns the default config directory
 func GetDefaultConfigDir() (dir string, err error) {
-	if runtime.GOOS == "windows" {
+	if env, ok := os.LookupEnv("PET_CONFIG_DIR"); ok {
+		dir = env
+	} else if runtime.GOOS == "windows" {
 		dir = os.Getenv("APPDATA")
 		if dir == "" {
 			dir = filepath.Join(os.Getenv("USERPROFILE"), "Application Data", "pet")
@@ -125,7 +130,7 @@ func GetDefaultConfigDir() (dir string, err error) {
 	} else {
 		dir = filepath.Join(os.Getenv("HOME"), ".config", "pet")
 	}
-	if err := os.MkdirAll(dir, 0700); err != nil {
+	if err := os.MkdirAll(dir, 0o700); err != nil {
 		return "", fmt.Errorf("cannot create directory: %v", err)
 	}
 	return dir, nil
